+++ eva/src/components/modalSettings.vue
@@ -1023,6 +1023,7 @@ export default {
       them: {},
       isConfirmModal: false,
       deleteMetricId: '',
+      isOsmServerChange: false,
     };
   },
   computed: {
@@ -1228,6 +1229,9 @@ export default {
           units: this.metricUnits,
         });
       }
+      if (this.element.startsWith('map')) {
+        this.changeSelectedLayer();
+      }
 
       const options = {
         ...this.options,
@@ -1248,6 +1252,25 @@ export default {
         this.deleteTheme();
       }
       this.cancelModal();
+    },
+    changeSelectedLayer() {
+      // Берем активную подложку из сторы
+      const selectedLayerFromStore = this.dashFromStore[this.element].options.selectedLayer;
+      // let newSelectedLayer = null;
+      if (!selectedLayerFromStore || selectedLayerFromStore?.name === 'Заданная в настройках') {
+        this.$store.commit('setState', [
+          {
+            object: this.dashFromStore[this.element].options,
+            prop: 'selectedLayer',
+            value: this.options.osmserver
+              ? {
+                name: 'Заданная в настройках',
+                tile: this.options.osmserver || [],
+              }
+              : null,
+          },
+        ]);
+      }
     },
     cancelModal() {
       this.active = false;