+++ CHANGELOG.md
@@ -5,10 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrealese]
+
+### Fixed
+- fixed change osmserver in map settings
+- fixed default osmserver and default search in dashMapUserSettings
+
 ## [1.86.1]
 
 ### Fixed
 - Fixed theme styles
+- fixed piechart visual inaccuracies
 
 
 ## [1.86.0]